feat: 修改 TelemetryBattery 结构体中的 Percentage 字段类型为 float64

+++ internal/domain/models.go
@@ -108,7 +108,7 @@ type TelemetryBattery struct {
 	RawADC       int    `json:"raw_adc"`
 	RawVoltageMV int    `json:"raw_voltage_mv"`
 	VoltageMV    int    `json:"voltage_mv"`
-	Percentage   int    `json:"percentage"`
+	Percentage   float64 `json:"percentage"`
 	Status       string `json:"status"`
 }
